pyproject.toml
[tool.poetry]
name        = "adhlal-etl"
version     = "0.1.2"
description = "ETL for Adhlal fashion-skills research"
authors     = ["Adhlal Data Team"]

[tool.poetry.dependencies]
python = "^3.11"
openai         = "^1.23.0"
pydantic       = "^2.7.0"
pandas         = "^2.2.2"
python-dotenv  = "^1.0.1"
psycopg2-binary= "^2.9.9"
tenacity       = "^8.2.3"
email-validator = "^2.0.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.4.2"
mypy = "^1.10.0"

[[tool.poetry.packages]]
include = "etl"
from = "src"

[tool.poetry.scripts]
adhlal-etl = "etl.__main__:main"

[build-system]
requires = ["poetry-core>=1.8.2"]
build-backend = "poetry.core.masonry.api" 